+++ src/transformer/rt-helper.ts
@@ -6,22 +6,34 @@ import {serialize, serializeExpression} from './serialize';
 const __RΦ = {
     m: (k, v) => (t, ...a) => t && Reflect.metadata ? Reflect.metadata(k, v)(t, ...a) : void 0,
     f: (f, d, n) => (d.forEach(d => d(f)), Object.defineProperty(f, "name", { value: n, writable: false }), f),
+    c: (c, d, dp, dsp, n) => (
+        d.forEach(d => d(c)),
+        dp.forEach(([p, d]) => d(c.prototype, p)),
+        dsp.forEach(([p, d]) => d(c, p)),
+        n ? Object.defineProperty(c, "name", { value: n, writable: false }) : undefined,
+        c
+    ),
     r: (o, a) => (Object.assign(o, a)),
     a: id => {
-            let t = __RΦ.t[id];
-            if (t.RΦ) {
-                let r = t.RΦ;
-                delete t.RΦ;
-                __RΦ.r(t, r(t));
-            }
-            else if (t.LΦ) {
-                let l = t.LΦ();
-                delete t.LΦ;
-                __RΦ.t[id] = t = l;
-            }
-            return t;
-    }
-}
+        let t = __RΦ.t[id];
+        if (t === void 0)
+            return void 0;
+        if (t.RΦ) {
+            let r = t.RΦ;
+            delete t.RΦ;
+            __RΦ.r(t, r(t));
+        }
+        else if (t.LΦ) {
+            let l = t.LΦ();
+            delete t.LΦ;
+            __RΦ.t[id] = t = l;
+        }
+        return t;
+    },
+    t: { [79]: { RΦ: t => ({ TΦ: "F", r: __RΦ.a(81), p: [], f: "" }) }, [81]: { TΦ: "5", name: "A" } }
+};
+
+
  */
 
 export function rtStore(typeMap: Map<number|string, ts.Expression>) {